requirements.txt
Django
djangorestframework
djoser
djangorestframework-simplejwt
python-dotenv
django-cors-headers
psycopg2-binary
stripe
numpy
scipy
scikit-learn
pytest
pytest-django
